chore(team): split step lifecycle manager tests
